# go/docs/api/openapi-ru.yaml
openapi: 3.0.3
info:
  title: BizOps360 Go API
  description: |
    Полная спецификация API для платформы BizOps360 Go Backend.
    
    **Всего эндпоинтов:** 15
    
    **Базовая информация:**
    - Версия API: 1.0.0
    - Базовый URL (DEV): https://bizops360-api-go-dev-gqqr4r256q-uc.a.run.app
    - Формат данных: JSON
    - Кодировка: UTF-8
    
    **Аутентификация:**
    - Некоторые эндпоинты требуют API ключ в заголовке `X-Api-Key`
    - Эндпоинты здоровья, календаря, Zapier и V1 не требуют аутентификации
    
    **Ограничения:**
    - Rate limiting: 100 запросов в минуту на IP
    - Максимальный размер запроса: 1MB
  version: 1.0.0
  contact:
    name: BizOps360 Support
    url: https://github.com/bizops360/bizops360
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки (Local Development)
  - url: https://bizops360-api-go-dev-gqqr4r256q-uc.a.run.app
    description: Сервер разработки (Development)
  - url: https://bizops360-api-go-prod-XXXXX-uc.a.run.app
    description: Сервер продакшена (Production)
...
paths:
...
  /api/stripe/deposit:
    post:
      tags:
        - Stripe
      summary: Создание депозита
      description: Генерация инвойса для депозита бронирования
      operationId: createDeposit
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Успешное создание депозита
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
  /api/estimate:
    post:
      tags:
        - Расчет стоимости
      summary: Расчет стоимости мероприятия
      description: Расчет стоимости мероприятия на основе даты, длительности и количества помощников
      operationId: calculateEstimate
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstimateRequest'
      responses:
        '200':
          description: Успешный расчет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstimateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API ключ для аутентификации
  schemas:
...
    DepositRequest:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
          description: Email клиента
          example: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        name:
          type: string
          description: Имя клиента
          example: "John Doe"
        estimatedTotal:
          type: number
          format: float
          description: Оценка общей стоимости
          example: 1000.0
        depositValue:
          type: number
          format: float
          nullable: true
          description: Значение депозита
          example: null
        deposit:
          type: number
          format: float
          nullable: true
          description: Депозит (альтернативное поле)
          example: null
        helpersCount:
          type: integer
          description: Количество помощников
          example: 2
        hours:
          type: number
          format: float
          description: Количество часов
          example: 4.0
        useTest:
          type: boolean
          default: false
          description: Использовать тестовый режим Stripe
          example: false
        dryRun:
          type: boolean
          default: false
          description: Тестовый запуск без реальных операций
          example: true
        mockStripe:
          type: boolean
          default: false
          description: Использовать мок Stripe
          example: false
      example:
        email: "bizops-dev-alexey-at-shevelyov-dot-com@shevelyov.com"
        name: "John Doe"
        estimatedTotal: 1000.0
        depositValue: null
        deposit: null
        helpersCount: 2
        hours: 4.0
        useTest: false
        dryRun: true
        mockStripe: false
    DepositResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        deposit:
          type: object
          properties:
            value:
              type: integer
              description: Значение в центах
            valueDollars:
              type: number
              format: float
              description: Значение в долларах
            percentage:
              type: number
              format: float
              description: Процент от общей стоимости
...
    EstimateRequest:
      type: object
      required:
        - eventDate
        - durationHours
        - numHelpers
      properties:
        eventDate:
          type: string
          format: date
          description: Дата мероприятия (YYYY-MM-DD)
          example: "2025-06-15"
        durationHours:
          type: number
          format: float
          description: Длительность в часах
          minimum: 0.1
          example: 4.0
        numHelpers:
          type: integer
          description: Количество помощников
          minimum: 1
          example: 2
      example:
        eventDate: "2025-06-15"
        durationHours: 4.0
        numHelpers: 2
    EstimateResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
          properties:
            year:
              type: integer
              description: Год мероприятия
            eventDate:
              type: string
              description: Дата мероприятия
            dateKey:
              type: string
              description: Ключ даты (YYYY-MM-DD)
            numHelpers:
              type: integer
              description: Количество помощников
            durationHours:
              type: number
              format: float
              description: Длительность в часах
            basePerHelper:
              type: number
              format: float
              description: Базовая ставка за помощника
            extraPerHourPerHelper:
              type: number
              format: float
              description: Дополнительная ставка за час за помощника
            baseSubtotal:
              type: number
              format: float
              description: Базовая подытог
            extraSubtotal:
              type: number
              format: float
              description: Дополнительная подытог
            subtotalBeforeAdjustments:
              type: number
              format: float
              description: Подытог до корректировок
            isSpecialDate:
              type: boolean
              description: Является ли дата специальной (праздник/повышенный спрос)
            specialLabel:
              type: string
              nullable: true
              description: Метка специальной даты
            rateType:
              type: string
              nullable: true
              description: Тип ставки
            specialDateMultiplier:
              type: number
              format: float
              nullable: true
              description: Множитель для специальной даты
            specialDateFlatIncrease:
              type: number
              format: float
              nullable: true
              description: Фиксированное увеличение для специальной даты
            totalCost:
              type: number
              format: float
              description: Общая стоимость
            currency:
              type: string
              description: Валюта
            breakdown:
              type: object
              description: Детальная разбивка стоимости
            calculationSummary:
              type: string
              description: Краткое описание расчета
            deposit:
              type: object
              description: Информация о депозите
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid JSON: unexpected end of JSON input"
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized: API key required"
    MethodNotAllowed:
      description: Метод не разрешен
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "method not allowed"